pyproject.toml
[tool.poetry]
name = "airbnb-clone2-backend"
version = "0.1.0"
description = ""
authors = ["June Roh <texasroh@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "airbnb_clone2_backend"}]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.1"
Pillow = "^9.2.0"
djangorestframework = "^3.13.1"
strawberry-graphql = {extras = ["debug-server"], version = "^0.133.5"}
strawberry-graphql-django = "^0.5.3"
PyJWT = "^2.5.0"
django-environ = "^0.9.0"
django-cors-headers = "^3.13.0"
requests = "^2.28.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
